+++ sphinxcontrib/bibtex/cache.py
@@ -158,6 +158,10 @@ class _FilterVisitor(ast.NodeVisitor):
     def visit_Str(self, node):
         return node.s
 
+    # NameConstant is Python 3.4 only so do not insist on coverage
+    def visit_NameConstant(self, node):  # pragma: no cover
+        return node.value
+
     def generic_visit(self, node):
         _raise_invalid_node(node)