+++ CHANGELOG.rst
@@ -21,6 +21,9 @@
 
   See documentation for details.
 
+* Multiple comma-separated citation keys per cite command (see issue
+  #61, suggested by Boris Kheyfets).
+
 * Support pypy.
 
 * Drop support for Python 2.6 and Python 3.2.